third_party/blink/web_tests/wpt_internal/css/css-anchor-position/inset-area-align-justify-expected.txt

This is a testharness.js-based test.
[FAIL] Offsets for inset-area: all
  assert_equals: Check expected offsetLeft expected 145 but got 10
[FAIL] Offsets for inset-area: top / center
  assert_equals: Check expected offsetLeft expected 145 but got 110
[FAIL] Offsets for inset-area: center / left
  assert_equals: Check expected offsetTop expected 145 but got 110
[FAIL] Offsets for inset-area: center / center
  assert_equals: Check expected offsetLeft expected 145 but got 110
[FAIL] Offsets for inset-area: center / right
  assert_equals: Check expected offsetTop expected 145 but got 110
[FAIL] Offsets for inset-area: bottom / center
  assert_equals: Check expected offsetLeft expected 145 but got 110
Harness: the test ran to completion.
